requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
apipkg==1.5
appdirs==1.4.4
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
behave==1.2.6
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
execnet==1.8.0
flake8==3.8.4
Flask==1.1.2
glob2==0.7
idna==2.10
iniconfig==1.1.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
nose==1.3.7
packaging==20.9
parameterized==0.7.4
parse==1.18.0
parse-type==0.5.2
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.1
pytest-forked==1.3.0
pytest-xdist==2.2.1
regex==2020.11.13
requests==2.25.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
